fix(protocol): widen frame header frameOff to u32 (#46)
Widen frameOff from u16 to u32: offsets within a 1 MiB block reach ~1 MiB, so they wrap past u16's 64 KiB bound and falsely signal a fresh block mid-block, corrupting reassembly. The 16-byte header is preserved by consuming the trailing reserved field, updated in lockstep across the TS and Go codecs and the shared AEAD vector.

+++ packages/wire/transfer_receiver_test.go
@@ -59,7 +59,7 @@ func (s *senderFrames) blockData(data []byte, blockSize, frameSize int) {
 			}
 			s.push(FrameHeaderInput{
 				Version: FrameVersion, Type: FrameBlockData, Flags: flags,
-				BlockIdx: uint32(blk), FrameOff: uint16(off),
+				BlockIdx: uint32(blk), FrameOff: uint32(off),
 			}, frag)
 		}
 		sum := sha256.Sum256(block)
